Bug fix and puzzle improvment.
Fixed a bug when if the start/end date cookie was missing, the server would result in a 502 error
Improved the JS puzzle. Now it computes a unique string by using client's signature as a key and yesterday's date as a seed and using XOR on these 2 values.
Also made a dynamic function on the server which calculates correct answer for each user.
Moved 'JavascriptPuzzleVars' so that it can use 'answer' variable (client's ip+signature string)
Made a comment about a need to improve random variables generation mechanism to exlude repeating variable names, which could lead to a script breaking.

I may missed something in this description, correct me if im wrong somewhere/if you disagree with provided changes.